refactor: icons and breadcrumbs
