+++ frontend/src/components/Primitives/Breadcrumb/Breadcrumb.tsx
@@ -1,9 +1,43 @@
 import Link from 'next/link';
 import React, { Fragment } from 'react';
 
+import { styled } from '@/styles/stitches/stitches.config';
 import Icon from '@/components/Primitives/Icons/Icon/Icon';
 import { BreadcrumbItemType, BreadcrumbType } from '@/types/board/Breadcrumb';
-import { StyledList, StyledBreadcrumbItem } from './styles';
+
+const StyledList = styled('ul', {
+  display: 'flex',
+  alignItems: 'center',
+  gap: 5,
+
+  margin: 0,
+  padding: 0,
+
+  listStyle: 'none',
+});
+
+const StyledBreadcrumbItem = styled('li', {
+  fontSize: '$14',
+  variants: {
+    isActive: {
+      true: {
+        color: '$primary800',
+      },
+      false: {
+        color: '$primary300',
+      },
+    },
+  },
+
+  a: {
+    color: '$primary300',
+    textDecoration: 'none',
+
+    '&:hover': {
+      color: '$primary800',
+    },
+  },
+});
 
 export type BreadcrumbProps = {
   items: BreadcrumbType;
