+++ frontend/src/components/Primitives/Icons/Svg/Svg.stories.tsx
@@ -5,7 +5,7 @@ import dedent from 'ts-dedent';
 
 import Svg from '@/components/Primitives/Icons/Svg/Svg';
 import Flex from '@/components/Primitives/Layout/Flex/Flex';
-import { SvgSizeType } from '../../../../stories/types/PrimitiveTypes';
+import { SvgSizeType } from '@/stories/types/PrimitiveTypes';
 
 const SIZE_OPTIONS: SvgSizeType[] = [12, 16, 18, 20, 24, 32, 40, 48, 100];
